+++ odoo/osv/osv.py
@@ -2,7 +2,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from ..exceptions import except_orm
-from .orm import Model, TransientModel, AbstractModel
+from ..models import Model, TransientModel, AbstractModel
 
 # Deprecated, kept for backward compatibility.
 # openerp.exceptions.Warning should be used instead.
